Changed omega to sum total user traffic observed by the attacker, so the attacker should lean heavily on the function W returning omega in their listen or attack decision. The total number of successful attacks by the attacker is now tracked as attacker_reward and is used as the attacker fitness function currently.
Moved defender decision to the last phase of the game since it's reactive to
either the user or attacker behavior. If we want to have it at the beginning we
could move it there (and adjust the decision to act on the previous round
affecting the current one), but to me it seems more intuitive to have it all
happen in order within a round.

Added config file parsing for the tunable game parameters.